+++ modules/network-annotation-processor/src/integrationTest/resources/org/apache/ignite/network/messages/internal/processor/InvalidGetterMessage.java
@@ -18,7 +18,7 @@
 package org.apache.ignite.network.messages.internal.processor;
 
 import org.apache.ignite.network.NetworkMessage;
-import org.apache.ignite.network.messages.annotations.AutoSerializable;
+import org.apache.ignite.network.processor.annotations.AutoSerializable;
 
 @AutoSerializable(messageFactory = AllTypesMessageFactory.class)
 public interface InvalidGetterMessage extends NetworkMessage {
