+++ modules/network-annotation-processor/src/integrationTest/resources/org/apache/ignite/network/messages/internal/processor/MissingBuilderMessage.java
@@ -19,7 +19,7 @@ package org.apache.ignite.network.messages.internal.processor;
 
 import java.util.Collection;
 import org.apache.ignite.network.NetworkMessage;
-import org.apache.ignite.network.messages.annotations.AutoSerializable;
+import org.apache.ignite.network.processor.annotations.AutoSerializable;
 
 @AutoSerializable(messageFactory = AllTypesMessageFactory.class)
 public interface MissingBuilderMessage extends NetworkMessage {
