+++ modules/network-annotation-processor/src/main/java/org/apache/ignite/network/processor/internal/MessageReaderMethodResolver.java
@@ -15,25 +15,22 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.network.messages.internal.processor;
+package org.apache.ignite.network.processor.internal;
 
-import java.util.Collection;
 import java.util.List;
-import java.util.Map;
 import javax.annotation.processing.ProcessingEnvironment;
 import javax.lang.model.element.ExecutableElement;
 import javax.lang.model.type.ArrayType;
 import javax.lang.model.type.DeclaredType;
-import javax.lang.model.type.TypeKind;
 import javax.lang.model.type.TypeMirror;
 import com.squareup.javapoet.CodeBlock;
-import org.apache.ignite.network.serialization.MessageWriter;
+import org.apache.ignite.network.serialization.MessageReader;
 import org.apache.ignite.plugin.extensions.communication.MessageCollectionItemType;
 
 /**
- * Class for resolving {@link MessageWriter} "write*" methods for the corresponding message field type.
+ * Class for resolving {@link MessageReader} "read*" methods for the corresponding message field type.
  */
-class MessageWriterMethodResolver {
+class MessageReaderMethodResolver {
     /** */
     private final BaseMethodNameResolver methodNameResolver;
 
@@ -41,66 +38,63 @@ class MessageWriterMethodResolver {
     private final MessageCollectionItemTypeConverter typeConverter;
 
     /** */
-    MessageWriterMethodResolver(ProcessingEnvironment processingEnvironment) {
+    MessageReaderMethodResolver(ProcessingEnvironment processingEnvironment) {
         methodNameResolver = new BaseMethodNameResolver(processingEnvironment);
         typeConverter = new MessageCollectionItemTypeConverter(processingEnvironment);
     }
 
     /**
-     * Resolves the "write" method by the type of the given message's getter method.
+     * Resolves the "read" method by the type of the given message's builder method.
      */
-    CodeBlock resolveWriteMethod(ExecutableElement getter) {
-        TypeMirror getterReturnType = getter.getReturnType();
-
-        if (getterReturnType.getKind() == TypeKind.VOID) {
-            throw new ProcessingException(String.format("Getter method %s does not return any value", getter));
+    CodeBlock resolveReadMethod(ExecutableElement builderSetter) {
+        if (builderSetter.getParameters().size() != 1) {
+            throw new ProcessingException("Invalid number of parameters of a Builder setter (expected 1): " + builderSetter);
         }
 
-        String parameterName = getter.getSimpleName().toString();
+        TypeMirror parameterType = builderSetter.getParameters().get(0).asType();
 
-        String methodName = methodNameResolver.resolveBaseMethodName(getterReturnType);
+        String parameterName = builderSetter.getSimpleName().toString();
+
+        String methodName = methodNameResolver.resolveBaseMethodName(parameterType);
 
         switch (methodName) {
             case "ObjectArray":
-                return resolveWriteObjectArray((ArrayType) getterReturnType, parameterName);
+                return resolveReadObjectArray((ArrayType) parameterType, parameterName);
             case "Collection":
-                return resolveWriteCollection((DeclaredType) getterReturnType, parameterName);
+                return resolveReadCollection((DeclaredType) parameterType, parameterName);
             case "Map":
-                return resolveWriteMap((DeclaredType) getterReturnType, parameterName);
+                return resolveReadMap((DeclaredType) parameterType, parameterName);
             default:
-                return CodeBlock.builder()
-                    .add("write$L($S, message.$L())", methodName, parameterName, parameterName)
-                    .build();
+                return CodeBlock.builder().add("read$L($S)", methodName, parameterName).build();
         }
     }
 
     /**
-     * Creates a {@link MessageWriter#writeObjectArray(String, Object[], MessageCollectionItemType)} method call.
+     * Creates a {@link MessageReader#readObjectArray(String, MessageCollectionItemType, Class)} method call.
      */
-    private CodeBlock resolveWriteObjectArray(ArrayType parameterType, String parameterName) {
+    private CodeBlock resolveReadObjectArray(ArrayType parameterType, String parameterName) {
         TypeMirror componentType = parameterType.getComponentType();
 
         return CodeBlock.builder()
             .add(
-                "writeObjectArray($S, message.$L(), $T.$L)",
-                parameterName,
+                "readObjectArray($S, $T.$L, $T.class)",
                 parameterName,
                 MessageCollectionItemType.class,
-                typeConverter.fromTypeMirror(componentType)
+                typeConverter.fromTypeMirror(componentType),
+                componentType
             )
             .build();
     }
 
     /**
-     * Creates a {@link MessageWriter#writeCollection(String, Collection, MessageCollectionItemType)} method call.
+     * Creates a {@link MessageReader#readCollection(String, MessageCollectionItemType)} method call.
      */
-    private CodeBlock resolveWriteCollection(DeclaredType parameterType, String parameterName) {
+    private CodeBlock resolveReadCollection(DeclaredType parameterType, String parameterName) {
         TypeMirror collectionGenericType = parameterType.getTypeArguments().get(0);
 
         return CodeBlock.builder()
             .add(
-                "writeCollection($S, message.$L(), $T.$L)",
-                parameterName,
+                "readCollection($S, $T.$L)",
                 parameterName,
                 MessageCollectionItemType.class,
                 typeConverter.fromTypeMirror(collectionGenericType)
@@ -109,10 +103,10 @@ class MessageWriterMethodResolver {
     }
 
     /**
-     * Creates a {@link MessageWriter#writeMap(String, Map, MessageCollectionItemType, MessageCollectionItemType)}
+     * Creates a {@link MessageReader#readMap(String, MessageCollectionItemType, MessageCollectionItemType, boolean)}
      * method call.
      */
-    private CodeBlock resolveWriteMap(DeclaredType parameterType, String parameterName) {
+    private CodeBlock resolveReadMap(DeclaredType parameterType, String parameterName) {
         List<? extends TypeMirror> typeArguments = parameterType.getTypeArguments();
 
         MessageCollectionItemType mapKeyType = typeConverter.fromTypeMirror(typeArguments.get(0));
@@ -120,8 +114,7 @@ class MessageWriterMethodResolver {
 
         return CodeBlock.builder()
             .add(
-                "writeMap($S, message.$L(), $T.$L, $T.$L)",
-                parameterName,
+                "readMap($S, $T.$L, $T.$L, false)",
                 parameterName,
                 MessageCollectionItemType.class,
                 mapKeyType,
