+++ modules/network-annotation-processor/src/main/java/org/apache/ignite/network/processor/internal/SerializationFactoryGenerator.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.network.messages.internal.processor;
+package org.apache.ignite.network.processor.internal;
 
 import javax.lang.model.element.Modifier;
 import javax.lang.model.element.TypeElement;
