+++ modules/network-annotation-processor/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -1,1 +1,1 @@
-org.apache.ignite.network.messages.internal.processor.AutoSerializableProcessor
+org.apache.ignite.network.processor.internal.AutoSerializableProcessor
